- Change the ATM stack functions to use intptr_t instead of int for opaque   arguments. - Fix a few other places that assumed that sizeof(int) == sizeof(void *).
Reviewed by:	mdodd


Former-commit-id: 65cf6b64da67be9363dd52fec64af8eb0f9a12f5

+++ sys/netatm/port.h
@@ -421,7 +421,7 @@ name|bfr
 parameter_list|,
 name|n
 parameter_list|)
-value|{				\ 	(bfr)->m_len = (n);				\ 	if ((bfr)->m_flags& M_EXT)			\ 		(bfr)->m_data = (caddr_t)(((u_int)(bfr)->m_ext.ext_buf	\ 			+ (bfr)->m_ext.ext_size - (n))& ~(sizeof(long) - 1));\ 	else						\ 		(bfr)->m_data = (caddr_t)(((u_int)(bfr)->m_dat + MLEN - (n)) \& ~(sizeof(long) - 1));		\ }
+value|{				\ 	(bfr)->m_len = (n);				\ 	if ((bfr)->m_flags& M_EXT)			\ 		(bfr)->m_data = (caddr_t)(((uintptr_t)(bfr)->m_ext.ext_buf \ 			+ (bfr)->m_ext.ext_size - (n))& ~(sizeof(long) - 1));\ 	else						\ 		(bfr)->m_data = (caddr_t)(((uintptr_t)(bfr)->m_dat + MLEN - (n)) \& ~(sizeof(long) - 1));		\ }
 end_define
 
 begin_define
